Update the state's name
