+++ src/app/components/countriesCard/fetchCountriesData.tsx
@@ -4,6 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import CountryCard from "./Card";
 import { useState } from 'react';
 import { ChevronDown } from 'lucide-react';
+import { RegionList } from './RegionList';
 
 export interface Country {
     cca3: string,
@@ -26,7 +27,7 @@ export interface Country {
   }
 
    export default function CountriesInfo() {
-    const [show , setShow] = useState(true)
+    const [showRegionList , setShowRegionList] = useState(false)
     const [selectedRegion, setSelectedRegion] = useState<string | undefined>(undefined);
 
     const { data } = useQuery<Country[]>({
@@ -40,13 +41,13 @@ export interface Country {
         <>
         <div className=' flex flex-col gap-48 items-end'>
           <div className=' fixed top-28 right-48 z-50'>
-            <div className="  flex justify-between items-center bg-primary-foreground shadow-md w-48 px-4 h-9 cursor-pointer  rounded-sm" onClick={()=> setShow(!show)}>
+            <div className="  flex justify-between items-center bg-primary-foreground shadow-md w-48 px-4 h-9 cursor-pointer  rounded-sm" onClick={()=> setShowRegionList(!showRegionList)}>
                 <p className=' text-primary'>Filter by region</p>
                 <ChevronDown className=' text-primary'/>
             </div>
-            {show && <RegionList setSelectedRegion = {setSelectedRegion}/>}
+            {showRegionList && <RegionList setSelectedRegion = {setSelectedRegion}/>}
           </div>
-          <div className=' flex text-black' onClick={()=> setShow(false)}>
+          <div className=' flex text-black' onClick={()=> setShowRegionList(false)}>
             <div className=' flex gap-16 flex-wrap items-center justify-center'>
               {data?.map((country) => (
                     <CountryCard 
@@ -67,20 +68,4 @@ export interface Country {
         
       );
   }
-  export const RegionList  = (props) =>{
-
-    const continents = ['All', 'Africa', 'Americas', 'Asia', 'Europe', 'Oceania'];
-        
-    return(
-           <div className=" flex flex-col items-start mt-2 bg-primary-foreground  shadow-md rounded-sm absolute z-50 w-48 ">
-                {continents.map((continent) => (
-                    <button className="hover:bg-slate-600 hover:text-white w-full text-start pl-4 py-2 text-primary"
-                        key={continent}
-                        onClick={() => props.setSelectedRegion(continent === 'All' ? undefined : continent.toLowerCase())}
-                    >
-                        {continent}
-                    </button>
-                ))}
-           </div>
-       )
- }
+
